Voeg bbcodefield toe aan invoervelden
