+++ lib/view/formulier/InvoerVelden.class.php
@@ -19,10 +19,12 @@ require_once 'view/formulier/invoervelden/LidField.class.php';
 require_once 'view/formulier/invoervelden/RechtenField.class.php';
 require_once 'view/formulier/invoervelden/StudieField.class.php';
 require_once 'view/formulier/invoervelden/TextareaField.class.php';
+require_once 'view/formulier/invoervelden/BBCodeField.php';
 require_once 'view/formulier/invoervelden/UrlField.class.php';
 require_once 'view/formulier/invoervelden/UsernameField.class.php';
 require_once 'view/formulier/invoervelden/WachtwoordField.class.php';
 require_once 'view/formulier/invoervelden/WachtwoordWijzigenField.class.php';
+require_once 'view/formulier/invoervelden/RequiredBBCodeField.class.php';
 require_once 'view/formulier/invoervelden/RequiredDuckField.class.php';
 require_once 'view/formulier/invoervelden/RequiredEmailField.class.php';
 require_once 'view/formulier/invoervelden/RequiredEntityField.class.php';
