+++ lib/view/formulier/invoervelden/RequiredBBCodeField.class.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * RequiredBBCodeField.class.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ */
+class RequiredBBCodeField extends BBCodeField {
+
+	public $required = true;
+
+}
